move token credential
