+++ Libraries/Microsoft.Teams.Api/Auth/TokenCredentials.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+
 using Microsoft.Teams.Common.Http;
 
 namespace Microsoft.Teams.Api.Auth;
