+++ Libraries/Microsoft.Teams.Api/Clients/BotTokenClient.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using Microsoft.Teams.Api.Auth;
 using Microsoft.Teams.Common.Http;
 
 using IHttpClientFactory = Microsoft.Teams.Common.Http.IHttpClientFactory;
