+++ Libraries/Microsoft.Teams.Apps/AppOptions.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using Microsoft.Teams.Api.Auth;
 using Microsoft.Teams.Apps.Plugins;
 
 namespace Microsoft.Teams.Apps;
@@ -12,7 +13,7 @@ public class AppOptions
     public Common.Storage.IStorage<string, object>? Storage { get; set; }
     public Common.Http.IHttpClient? Client { get; set; }
     public Common.Http.IHttpClientFactory? ClientFactory { get; set; }
-    public Common.Http.IHttpCredentials? Credentials { get; set; }
+    public IHttpCredentials? Credentials { get; set; }
     public IList<IPlugin> Plugins { get; set; } = [];
     public OAuthSettings OAuth { get; set; } = new OAuthSettings();
