--- Libraries/Microsoft.Teams.Common/Http/HttpCredentialsFactory.cs
@@ -1,10 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-namespace Microsoft.Teams.Common.Http;
-
-public interface IHttpCredentialsFactory
-{
-    public IHttpCredentials? GetCredentials();
-    public Task<IHttpCredentials?> GetCredentialsAsync();
-}
